lab_1/requirements.txt
kafka-python==2.3.0
pandas
numpy
joblib
scikit-learn
pyarrow
streamlit
streamlit_autorefresh
